+++ ubo_app/services/030-ip/ubo_handle.py
@@ -6,13 +6,15 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from ubo_handle import ReducerRegistrar, register
 
+    from ubo_app.utils.types import Subscriptions
 
-def setup(register_reducer: ReducerRegistrar) -> None:
+
+async def setup(register_reducer: ReducerRegistrar) -> Subscriptions:
     from reducer import reducer
     from setup import init_service
 
     register_reducer(reducer)
-    init_service()
+    return await init_service()
 
 
 register(